reddit-affiliate-bot/requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
Flask==3.0.0
